[MISC] Copy type_list submodule to utility.
